🔧 refactor: Consolidate imports and enhance matcher functionality
- Merged multiple import statements for better organization and readability across various matchers.
- Introduced a new `helper.py` module to centralize utility functions, improving code reuse and maintainability.
- Updated matcher implementations to utilize the new helper functions, streamlining the code and enhancing clarity.
- Added additional metadata to the plugin configuration for improved documentation and author information.
