+++ nonebot_plugin_resolver2/matchers/kugou.py
@@ -1,8 +1,7 @@
 import re
 
+from nonebot import logger, on_message
 from nonebot.adapters.onebot.v11 import MessageSegment
-from nonebot.log import logger
-from nonebot.plugin.on import on_message
 
 from nonebot_plugin_resolver2.config import NEED_UPLOAD, NICKNAME
 from nonebot_plugin_resolver2.download import download_audio
@@ -10,8 +9,8 @@ from nonebot_plugin_resolver2.download.utils import keep_zh_en_num
 from nonebot_plugin_resolver2.parsers.kugou import KuGou
 
 from .filter import is_not_in_disabled_groups
+from .helper import get_file_seg
 from .preprocess import ExtractText, r_keywords
-from .utils import get_file_seg
 
 kugou = on_message(rule=is_not_in_disabled_groups & r_keywords("kugou.com"))
 kugou_parser = KuGou()
