+++ nonebot_plugin_resolver2/matchers/twitter.py
@@ -2,18 +2,17 @@ import re
 from typing import Any
 
 import aiohttp
-from nonebot import on_keyword
+from nonebot import logger, on_keyword
 from nonebot.adapters.onebot.v11 import MessageEvent, MessageSegment
-from nonebot.log import logger
 from nonebot.rule import Rule
 
 from nonebot_plugin_resolver2.config import NICKNAME, PROXY
 from nonebot_plugin_resolver2.constant import COMMON_HEADER
-from nonebot_plugin_resolver2.download import download_img
+from nonebot_plugin_resolver2.download import download_img, download_video
 from nonebot_plugin_resolver2.parsers.base import ParseException
 
 from .filter import is_not_in_disabled_groups
-from .utils import get_video_seg
+from .helper import get_video_seg
 
 twitter = on_keyword(keywords={"x.com"}, rule=Rule(is_not_in_disabled_groups))
 
@@ -35,7 +34,8 @@ async def _(event: MessageEvent):
     except ParseException as e:
         await twitter.finish(f"{NICKNAME}解析 | 小蓝鸟 - {e}")
     if video_url:
-        await twitter.send(await get_video_seg(url=video_url, proxy=PROXY))
+        video_path = await download_video(url=video_url, proxy=PROXY)
+        await twitter.send(get_video_seg(video_path))
     if pic_url:
         img_path = await download_img(url=pic_url, proxy=PROXY)
         await twitter.send(MessageSegment.image(img_path))
