+++ nonebot_plugin_resolver2/matchers/xiaohongshu.py
@@ -1,8 +1,7 @@
 import re
 
+from nonebot import logger, on_message
 from nonebot.adapters.onebot.v11 import MessageSegment
-from nonebot.log import logger
-from nonebot.plugin.on import on_message
 
 from nonebot_plugin_resolver2.config import NICKNAME
 from nonebot_plugin_resolver2.download import download_imgs_without_raise, download_video
@@ -10,8 +9,8 @@ from nonebot_plugin_resolver2.parsers.base import ParseException
 from nonebot_plugin_resolver2.parsers.xiaohongshu import parse_url
 
 from .filter import is_not_in_disabled_groups
+from .helper import get_video_seg, send_segments
 from .preprocess import ExtractText, r_keywords
-from .utils import get_video_seg, send_segments
 
 xiaohongshu = on_message(rule=is_not_in_disabled_groups & r_keywords("xiaohongshu.com", "xhslink.com"))
